// components/common/types/disperse.types.ts
export type TAmountInputElement = {
	amount: string | undefined;
	normalizedBigAmount: TNormalizedBN;
	status: 'pending' | 'success' | 'error' | 'none';
	isValid: boolean | 'undetermined';
	error?: string | undefined;
};

/**********************************************************************************************
 ** TDisperseInput contains the receiver and the value of the disperse operation. The UUID is
 ** used to identify the input and control it from outside. It acts as a unique reference.
 *********************************************************************************************/
export type TDisperseInput = {
	receiver: TInputAddressLike;
	value: TAmountInputElement;
	UUID: string;
};

/**********************************************************************************************
 ** TDisperseConfiguration contains all the necessary information to perform a disperse
 ** operation
 *********************************************************************************************/
export type TDisperseConfiguration = {
	tokenToSend: TToken | undefined;
	inputs: TDisperseInput[];
};
// ...
export type TDisperseActions =
	| {type: 'SET_TOKEN_TO_SEND'; payload: TToken | undefined}
	| {type: 'SET_RECEIVERS'; payload: TDisperseInput[]}
	| {type: 'ADD_RECEIVERS'; payload: TDisperseInput[]}
	| {type: 'PASTE_RECEIVERS'; payload: TDisperseInput[]}
	| {type: 'DEL_RECEIVER_BY_UUID'; payload: string}
	| {type: 'SET_RECEIVER'; payload: Partial<TInputAddressLike> & {UUID: string}}
	| {type: 'SET_VALUE'; payload: Partial<TAmountInputElement> & {UUID: string}}
	| {type: 'CLEAR_RECEIVERS'; payload: undefined}
	| {type: 'RESET'; payload: undefined};
// ...
export type TDisperseContext = {
	configuration: TDisperseConfiguration;
	dispatchConfiguration: Dispatch<TDisperseActions>;
	isDispersed: boolean;
	onResetDisperse: () => void;
};